upload

+++ StudentManagement/src/main/java/tridm/StudentManagement/repositories/SinhVienRepository.java
@@ -4,7 +4,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import tridm.StudentManagement.models.SinhVien;
 
-public interface SinhVienRepository extends JpaRepository<SinhVien, Integer>{
+public interface SinhVienRepository extends JpaRepository<SinhVien, Long>{
 
     
 }
